+++ app/components/CronoSetsList.tsx/SetItem.tsx
@@ -1,8 +1,8 @@
 import React, { FC, useMemo } from 'react';
 import { StyleSheet } from 'react-native';
 import { COLOR_GREEN_NORMAL, COLOR_RED_NORMAL, FONT_COLOR_GREY } from '../../commonStyles';
-import { CronoSetType } from '../../crono/redux/CronoTypes';
-import { SetMode, SetType } from '../../editor/enums';
+import { CronoSetType } from '../../modules/crono/redux/CronoTypes';
+import { SetMode, SetType } from '../../modules/editor/enums';
 import generateTimestamp from '../../utils/time/generateTimestamp';
 import secondsToTimeString from '../../utils/time/secondsToTimeString';
 import TextComponent from '../TextComponent/OldTextComponent';
