+++ app/components/CronoSetsList.tsx/SetsList.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { CronoSetType } from '../../crono/redux/CronoTypes';
+import { CronoSetType } from '../../modules/crono/redux/CronoTypes';
 import SetItem from './SetItem';
 import { Grid } from './SetsList.styled';
