+++ app/modules/crono/pure/calculateAverageContractions.tsx
