+++ app/modules/crono/redux/actions/replaceSetAction.ts
@@ -1,5 +1,5 @@
 import { Action } from 'redux';
-import * as reduxActions from '../../../redux/actions';
+import * as reduxActions from '../../../../redux/actions';
 import { CronoSetType } from '../CronoTypes';
 
 type ReplaceSetAction = Action & {
