+++ app/modules/crono/redux/actions/setInitialStateAction.ts
@@ -1,5 +1,5 @@
 import { Action } from 'redux';
-import * as reduxActions from '../../../redux/actions';
+import * as reduxActions from '../../../../redux/actions';
 import { ImmutableJSCronoType } from '../CronoTypes';
 
 export type InitialStateAction = Action & {
