+++ app/modules/crono/redux/creators/updateContractionsAverageAction.ts
