+++ app/modules/editor/redux/actions/setEditorTableBase.ts
