+++ app/modules/editor/redux/actions/setEditorTableDuration.ts
