+++ app/modules/editor/redux/hooks.ts
@@ -1,6 +1,6 @@
 import { TypedUseSelectorHook, useSelector, useDispatch } from 'react-redux';
-import { RootState } from '../../redux/types';
-import { AppDispatch } from '../../../App';
+import { RootState } from '../../../redux/types';
+import { AppDispatch } from '../../../../App';
 
 export const useAppDispatch: () => AppDispatch = useDispatch;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
