+++ app/redux/reducers.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import cronoReducer from '../crono/redux/cronoReducer';
-import editorReducer from '../editor/redux/editorReducer';
+import cronoReducer from '../modules/crono/redux/cronoReducer';
+import editorReducer from '../modules/editor/redux/editorReducer';
 import { RootState } from './types';
 
 const initReducers = () => {
