+++ app/redux/types.ts
@@ -1,7 +1,7 @@
 import { AnyAction } from 'redux';
 import { ThunkAction } from 'redux-thunk';
-import { ImmutableJSCronoType } from '../crono/redux/CronoTypes';
-import { ImmutableJSEditorStateType } from '../editor/redux/editorTypes';
+import { ImmutableJSCronoType } from '../modules/crono/redux/CronoTypes';
+import { ImmutableJSEditorStateType } from '../modules/editor/redux/editorTypes';
 
 export type RootState = {
   editor: ImmutableJSEditorStateType;
