+++ app/routes/Crono/components/LiveCounter/LiveCounter.tsx
@@ -2,8 +2,8 @@ import React, { FC, useMemo } from 'react';
 import styled from 'styled-components/native';
 import { COLOR_GREEN_NORMAL, COLOR_RED_NORMAL, FONT_COLOR_GREY, FONT_SIZE } from '../../../../commonStyles';
 import InfoBlock from '../../../../components/InfoBlock';
-import { CronoSetType, CronoStateType } from '../../../../crono/redux/CronoTypes';
-import { SetMode, SetType, TableTypeEnum } from '../../../../editor/enums';
+import { CronoSetType, CronoStateType } from '../../../../modules/crono/redux/CronoTypes';
+import { SetMode, SetType, TableTypeEnum } from '../../../../modules/editor/enums';
 import generateTimestamp from '../../../../utils/time/generateTimestamp';
 
 const LiveCounterWrapper = styled.View`
