+++ app/routes/EnduranceEditor/components/EnduranceForm/EnduranceForm.tsx
@@ -2,12 +2,12 @@ import React, { FC, useEffect } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import CronoStartButton from '../../../../components/CronoStartButton/CronoStartButton';
-import * as editorActions from '../../../../editor/redux/editorActions';
+import * as editorActions from '../../../../modules/editor/redux/editorActions';
 import {
   EditorActionsTypes,
   ImmutableJSEditorSetType,
   ImmutableJSEditorType,
-} from '../../../../editor/redux/editorTypes';
+} from '../../../../modules/editor/redux/editorTypes';
 import { RootState } from '../../../../redux/types';
 import { Wrapper } from './EnduranceForm.styled';
 import EnduranceMainForm from './EnduranceMainForm';
