+++ app/routes/StaticEditor/components/StaticForm/StaticMainForm.tsx
@@ -3,8 +3,8 @@ import { COLOR_GREEN_NORMAL, COLOR_RED_NORMAL } from '../../../../commonStyles';
 import InfoBlock from '../../../../components/InfoBlock';
 import TableBaseInput from '../StaticFormInputs/TableBaseInput';
 import TableTypeInput from '../StaticFormInputs/TableTypeInput';
-import { TableTypeEnum } from '../../../../editor/enums';
-import { ImmutableJSEditorStateType } from '../../../../editor/redux/editorTypes';
+import { TableTypeEnum } from '../../../../modules/editor/enums';
+import { ImmutableJSEditorStateType } from '../../../../modules/editor/redux/editorTypes';
 import * as SC from './StaticForm.styled';
 
 const titleByType = {
