+++ app/routes/StaticEditor/components/StaticFormInputs/EditorTimerInput.tsx
@@ -3,11 +3,11 @@ import { StyleSheet } from 'react-native';
 import { COLOR_GREEN_NORMAL, COLOR_RED_NORMAL, FONT_COLOR_GREY } from '../../../../commonStyles';
 import LongTouchButton from '../../../../components/LongTouchButton';
 import TextComponent from '../../../../components/TextComponent/OldTextComponent';
-import { SetType } from '../../../../editor/enums';
+import { SetType } from '../../../../modules/editor/enums';
 import secondsToTimeString from '../../../../utils/time/secondsToTimeString';
 import * as SC from './EditorTimerInput.styled';
-import { useAppDispatch } from '../../../../editor/redux/hooks';
-import { decreaseTimeItem, increaseTimeItem } from '../../../../editor/redux/editorActions';
+import { useAppDispatch } from '../../../../modules/editor/redux/hooks';
+import { decreaseTimeItem, increaseTimeItem } from '../../../../modules/editor/redux/editorActions';
 
 type EditorTimerInputProps = {
   index: number;
